+++ github-finder/src/components/index.js
@@ -1,4 +1,5 @@
 export { default as HeaderWrapper } from "./common/HeaderWrapper";
 export { default as MainWrapper } from "./main/MainWrapper";
 export { default as SearchForm } from "./main/SearchForm";
-export { default as SearchResult } from "./main/SearchResult";
+export { default as SearchResult } from "./result/SearchResult";
+export { default as DescriptionWrapper } from "./result/DescriptionWrapper";
