+++ github-finder/src/components/result/githubIcon.png
